SpineModellling_python/requirements.txt

cycler==0.12.1
dsntnn==0.5.3

fonttools==4.60.1
fsspec==2025.9.0
kiwisolver==1.4.7


matplotlib==3.9.4
mkl-service==2.4.0

opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
Pillow==10.4.0
pydicom==2.4.4
pylibjpeg==2.0.1
pylibjpeg-libjpeg==2.3.0
pylibjpeg-openjpeg==2.5.0
gdcm==1.1
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0

python-dateutil==2.9.0.post0
pytz==2025.2

six==1.17.0
sympy==1.13.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1

tzdata==2025.2

pyqtgraph==0.13.7
PyOpenGL==3.1.10
PyOpenGL_accelerate==3.1.10
PyQt3D==5.15.7
numpy-stl==3.2.0
